fix(utils): cast tensors to float32 before torch.cdist to support BFloat16

+++ src/heretic/utils.py
@@ -241,9 +241,9 @@ def mean_distances_to_knn(a: Tensor, b: Tensor, k: int) -> Tensor:
     if b.shape[0] == 0:
         return torch.zeros(a.shape[0], device=a.device)
     k = min(k, b.shape[0])
-    distances = torch.cdist(a, b)
+    distances = torch.cdist(a.float(), b.float())
     nearest_distances, _ = distances.topk(k, dim=1, largest=False)
-    return nearest_distances.mean(1)
+    return nearest_distances.mean(1).to(a.dtype)
 
 
 def empty_cache():
